release: Chromium 18.0.1025 (18.0.1025.99)
Compare: https://github.com/chromium/chromium/compare/f731667b158756deeb5531229e57be2de56c5797...1b80f3a67261e81379d1990f1c5fc969ee2fb966

+++ net/base/default_origin_bound_cert_store.cc
@@ -1,4 +1,4 @@
-// Copyright (c) 2011 The Chromium Authors. All rights reserved.
+// Copyright (c) 2012 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -30,6 +30,8 @@ void DefaultOriginBoundCertStore::FlushStore(
 bool DefaultOriginBoundCertStore::GetOriginBoundCert(
     const std::string& origin,
     SSLClientCertType* type,
+    base::Time* creation_time,
+    base::Time* expiration_time,
     std::string* private_key_result,
     std::string* cert_result) {
   base::AutoLock autolock(lock_);
@@ -42,6 +44,8 @@ bool DefaultOriginBoundCertStore::GetOriginBoundCert(
 
   OriginBoundCert* cert = it->second;
   *type = cert->type();
+  *creation_time = cert->creation_time();
+  *expiration_time = cert->expiration_time();
   *private_key_result = cert->private_key();
   *cert_result = cert->cert();
 
@@ -51,6 +55,8 @@ bool DefaultOriginBoundCertStore::GetOriginBoundCert(
 void DefaultOriginBoundCertStore::SetOriginBoundCert(
     const std::string& origin,
     SSLClientCertType type,
+    base::Time creation_time,
+    base::Time expiration_time,
     const std::string& private_key,
     const std::string& cert) {
   base::AutoLock autolock(lock_);
@@ -58,7 +64,9 @@ void DefaultOriginBoundCertStore::SetOriginBoundCert(
 
   InternalDeleteOriginBoundCert(origin);
   InternalInsertOriginBoundCert(
-      origin, new OriginBoundCert(origin, type, private_key, cert));
+      origin,
+      new OriginBoundCert(
+          origin, type, creation_time, expiration_time, private_key, cert));
 }
 
 void DefaultOriginBoundCertStore::DeleteOriginBoundCert(
@@ -68,17 +76,28 @@ void DefaultOriginBoundCertStore::DeleteOriginBoundCert(
   InternalDeleteOriginBoundCert(origin);
 }
 
-void DefaultOriginBoundCertStore::DeleteAll() {
+void DefaultOriginBoundCertStore::DeleteAllCreatedBetween(
+    base::Time delete_begin,
+    base::Time delete_end) {
   base::AutoLock autolock(lock_);
   InitIfNecessary();
   for (OriginBoundCertMap::iterator it = origin_bound_certs_.begin();
-       it != origin_bound_certs_.end(); ++it) {
-    OriginBoundCert* cert = it->second;
-    if (store_)
-      store_->DeleteOriginBoundCert(*cert);
-    delete cert;
+       it != origin_bound_certs_.end();) {
+    OriginBoundCertMap::iterator cur = it;
+    ++it;
+    OriginBoundCert* cert = cur->second;
+    if ((delete_begin.is_null() || cert->creation_time() >= delete_begin) &&
+        (delete_end.is_null() || cert->creation_time() < delete_end)) {
+      if (store_)
+        store_->DeleteOriginBoundCert(*cert);
+      delete cert;
+      origin_bound_certs_.erase(cur);
+    }
   }
-  origin_bound_certs_.clear();
+}
+
+void DefaultOriginBoundCertStore::DeleteAll() {
+  DeleteAllCreatedBetween(base::Time(), base::Time());
 }
 
 void DefaultOriginBoundCertStore::GetAllOriginBoundCerts(
